requirements.txt
PyYAML==3.12
Pygments==2.1.3
decorator==4.0.11
ipython==5.1.0
ipython-genutils==0.1.0
oauthlib==2.0.1
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.9
ptyprocess==0.5.1
pyzmq==16.0.2
redis==2.10.5
requests==2.12.4
requests-oauthlib==0.7.0
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.1
tweepy==3.5.0
wcwidth==0.1.7
